requirements.txt
absl-py==2.3.1
aiodns==3.5.0
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiokafka==0.10.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
autopep8==2.3.2
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.34.162
botocore==1.34.162
cachelib==0.13.0
cachetools==4.2.2
ccxt==4.3.86
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
clickhouse-connect==0.8.18
clickhouse-driver==0.2.9
colorama==0.4.6
confluent-kafka==2.5.0
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.3
dash-extensions==2.0.4
dataclass-wizard==0.35.1
decorator==5.2.1
defusedxml==0.7.1
deptry==0.23.1
discord.py==2.5.2
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
EditorConfig==0.17.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
Flask-Caching==2.3.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
git-filter-repo==2.47.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
identify==2.6.13
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsbeautifier==1.15.4
jsonschema==4.22.0
jsonschema-specifications==2025.4.1
kafka-python==2.0.2
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.6.0
lz4==4.4.4
magic-filter==1.0.12
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.2.0
more-itertools==10.7.0
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
namex==0.1.0
narwhals==2.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
openai==1.98.0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.6
packaging==25.0
pandas==2.2.2
pandas-ta==0.3.14b0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycodestyle==2.14.0
pycparser==2.22
# ---- FIXED: single source of truth for Pydantic (compatible with FastAPI & aiogram) ----
pydantic>=2.10,<2.12
pydantic-settings==2.5.2
# ----------------------------------------------------------------------------------------
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
pyparsing==3.2.3
pyright==1.1.403
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-telegram-bot==13.15
pytz==2025.2
pywin32==306 ; sys_platform == "win32"
PyYAML==6.0.1
pyzmq==27.0.0
redis==5.0.7
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.13.0
retrying==1.4.2
rich==13.7.1
rich-toolkit==0.15.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.8
s3transfer==0.10.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.16.0
seaborn==0.13.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.37.2
sympy==1.14.0
ta==0.11.0
tenacity==8.3.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow==2.14.0 ; sys_platform != "win32"
tensorflow-intel==2.14.0 ; sys_platform == "win32"
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.0
vaderSentiment==3.3.2
virtualenv==20.34.0
vulture==2.14
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.14.1
xgboost==3.0.4
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
